HW1
Assignment description from my professor:

This assignment focuses on operator overloading but is non-prescriptive as to implementation. In other words, it’s up to you to design a class with which to demonstrate the required operator overloads.
Note: You will need a “Driver” class to demonstrate your code’s operations. A “Driver” class is “a program that does nothing but test a method.” In other words, you need a class with a main() that is used to test your created object and actions (e.g. operator overloads).